release: v0.1.8 (sidebar refactor, updated docs)
