+++ README.md
@@ -15,7 +15,7 @@
 
 * **Status:** Beta (`0.x`)
 * **License:** MIT
-* **Python:** 3.8–3.12
+* **Python:** 3.8–3.13
 * **OS:** Linux / macOS / Windows (incl. WSL)
 
 ---
